Move: SessionHandle to session module
